% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph_from_adj_mat_module.R
\name{build_graph_from_adj_mat_module}
\alias{build_graph_from_adj_mat_module}
\title{Build a graph object from a adjacency matrix and module Info}
\usage{
build_graph_from_adj_mat_module(
  adjacency_matrix,
  node_annotation = NULL,
  directed = F,
  top_modules = 15,
  seed = 1115
)
}
\arguments{
\item{adjacency_matrix}{Numeric matrix.
A numeric matrix with adjacency matrix.}

\item{node_annotation}{Data Frame
#' The annotation file of nodes in network contain \code{node} and \code{Modularity}}

\item{directed}{Logical (default: \code{FALSE}).
Whether edges between nodes are directed.}

\item{top_modules}{Integer.
Number of top-ranked modules to select.}

\item{seed}{Integer (default = 1115).
Random seed for reproducibility.}
}
\value{
An graph object representing the correlation network.
}
\description{
Build a graph object from a adjacency matrix and module Info
}
\examples{
NULL
}
