% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph_from_double_mat.R
\name{build_graph_from_double_mat}
\alias{build_graph_from_double_mat}
\title{Build a graph object from Two numeric matrix}
\usage{
build_graph_from_double_mat(
  mat1,
  mat2,
  module.method = c("Fast_greedy", "Walktrap", "Edge_betweenness", "Spinglass"),
  node_annotation = NULL,
  directed = F,
  top_modules = 15,
  seed = 1115
)
}
\arguments{
\item{mat1}{Numeric matrix.
A numeric matrix with samples in colums and variables in rows}

\item{mat2}{Numeric matrix.
A numeric matrix with samples in colums and variables in rows}

\item{module.method}{Character.
Network community detection (module identification) method.
Options include "Fast_greedy", "Walktrap", "Edge_betweenness", and "Spinglass".}

\item{node_annotation}{Data frame.
Optional node annotation table, containing metadata such as taxonomy or functional categories.}

\item{directed}{Logical (default: \code{FALSE}).
Whether edges between nodes are directed.}

\item{top_modules}{Integer.
Number of top-ranked modules to retain for downstream visualization or analysis.}

\item{seed}{Integer (default = 1115).
Random seed for reproducibility.}
}
\value{
A graph object representing the correlation-based two numeric matrix.
}
\description{
Build a graph object from Two numeric matrix
}
\examples{
\donttest{
set.seed(1)
mat1 <- matrix(stats::rnorm(15 * 20), nrow = 15, ncol = 20)
mat2 <- matrix(stats::rnorm(10 * 20), nrow = 10, ncol = 20)
rownames(mat1) <- paste0("A", seq_len(15))
rownames(mat2) <- paste0("B", seq_len(10))
colnames(mat1) <- colnames(mat2) <- paste0("sample", seq_len(20))
obj <- build_graph_from_double_mat(
  mat1          = mat1,
  mat2          = mat2,
  module.method = "Fast_greedy"
)
obj
}
}
