% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph_from_double_mat_with_module.R
\name{build_graph_from_double_mat_with_module}
\alias{build_graph_from_double_mat_with_module}
\title{Build a graph object from Two numeric matrix}
\usage{
build_graph_from_double_mat_with_module(
  mat1,
  mat2,
  node_annotation = NULL,
  directed = F,
  top_modules = 15,
  seed = 1115
)
}
\arguments{
\item{mat1}{Numeric matrix.
A numeric matrix with samples in colums and variables in rows}

\item{mat2}{Numeric matrix.
A numeric matrix with samples in colums and variables in rows}

\item{node_annotation}{Data frame.
Optional node annotation table, containing metadata such as taxonomy or functional categories.}

\item{directed}{Logical (default: \code{FALSE}).
Whether edges between nodes are directed.}

\item{top_modules}{Integer.
Number of top-ranked modules to retain for downstream visualization or analysis.}

\item{seed}{Integer (default = 1115).
Random seed for reproducibility.}
}
\value{
A graph object representing the correlation-based two numeric matrix.
}
\description{
Build a graph object from Two numeric matrix
}
\examples{
NULL
}
