% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph_from_enrichment.R
\name{build_graph_from_enrichGO}
\alias{build_graph_from_enrichGO}
\title{Build a  graph object from GO enrichment analysis}
\usage{
build_graph_from_enrichGO(df)
}
\arguments{
\item{df}{Data frame.
The output from GO enrichmen analysis in clusterProfiler}
}
\value{
An graph object with GO enrichment analysis
}
\description{
Build a  graph object from GO enrichment analysis
}
\examples{
\dontrun{
# Requires clusterProfiler enrichment result
ego <- clusterProfiler::enrichGO(gene, OrgDb = org.Hs.eg.db, ont = "ALL")
obj <- build_graph_from_enrichGO(df = as.data.frame(ego))
}
}
