% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph_from_igraph.R
\name{build_graph_from_igraph}
\alias{build_graph_from_igraph}
\title{Build a graph object from a graph object}
\usage{
build_graph_from_igraph(
  igraph,
  module.method = c("Fast_greedy", "Walktrap", "Edge_betweenness", "Spinglass"),
  node_annotation = NULL,
  top_modules = 15,
  seed = 1115
)
}
\arguments{
\item{igraph}{a igraph object}

\item{module.method}{Character.
Network community detection (module identification) method.
Options include "Fast_greedy", "Walktrap", "Edge_betweenness", and "Spinglass".}

\item{node_annotation}{Data Frame
The annotation file of nodes in network}

\item{top_modules}{Integer.
Number of top-ranked modules to select.}

\item{seed}{Integer (default = 1115).
Random seed for reproducibility.}
}
\value{
An graph object
}
\description{
Build a graph object from a graph object
}
\examples{
NULL
}
