% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph_from_multi_mat.R
\name{build_graph_from_multi_mat}
\alias{build_graph_from_multi_mat}
\title{Build a graph object from multiple numeric matrices}
\usage{
build_graph_from_multi_mat(
  mat1,
  mat2,
  ...,
  module.method = c("Fast_greedy", "Walktrap", "Edge_betweenness", "Spinglass"),
  node_annotation = NULL,
  directed = FALSE,
  top_modules = 15,
  seed = 1115
)
}
\arguments{
\item{mat1}{Numeric matrix.
A numeric matrix with samples in colums and variables in rows}

\item{mat2}{Numeric matrix.
A numeric matrix with samples in colums and variables in rows}

\item{...}{Additional numeric matrices with samples in columns and variables in rows.}

\item{module.method}{Character.
Network community detection (module identification) method.
Options include "Fast_greedy", "Walktrap", "Edge_betweenness", and "Spinglass".}

\item{node_annotation}{Data frame.
Optional node annotation table, containing metadata such as taxonomy or functional categories.}

\item{directed}{Logical (default: \code{FALSE}).
Whether edges between nodes are directed.}

\item{top_modules}{Integer.
Number of top-ranked modules to retain for downstream visualization or analysis.}

\item{seed}{Integer (default = 1115).
Random seed for reproducibility.}
}
\value{
A graph object representing the correlation-based multi numeric matrix.
}
\description{
Build a graph object from multiple numeric matrices
}
\examples{
NULL
}
