% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph_from_pie.R
\name{build_graph_from_pie}
\alias{build_graph_from_pie}
\title{Build a pie graph object from a data frame}
\usage{
build_graph_from_pie(df, node_annotation = NULL, directed = F, seed = 1115)
}
\arguments{
\item{df}{Data frame.
Edge list with columns \code{from}, \code{to}, and optionally \code{weight}.
If \code{weight} is absent, an unweighted graph is constructed.}

\item{node_annotation}{Data Frame
The annotation file of nodes in network}

\item{directed}{Logical (default: \code{FALSE}).
Whether edges between nodes are directed.}

\item{seed}{Integer (default = 1115).
Random seed for reproducibility.}
}
\value{
An graph object representing the correlation network.
  Node/edge attributes include correlation statistics and (optionally) module labels.
}
\description{
Build a pie graph object from a data frame
}
\examples{
data(ppi_example)
obj <- build_graph_from_pie(
  df              = ppi_example$ppi,
  node_annotation = ppi_example$annotation
)
obj
}
