% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layout_gephi.R
\name{create_layout_gephi}
\alias{create_layout_gephi}
\title{Create gephi layout}
\usage{
create_layout_gephi(graph_obj, node_add = 7, r = 0.1)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}

\item{node_add}{Integer (default = 7).
Number of nodes to add in each layer of the layout.}

\item{r}{Numeric (default = 1).
Radius increment for concentric or layered layouts.}
}
\value{
layout.
}
\description{
internal function
}
\examples{
NULL
}
\keyword{internal}
