% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layout_heart_centered.R
\name{create_layout_heart_centered}
\alias{create_layout_heart_centered}
\title{Create heart centered layout}
\usage{
create_layout_heart_centered(
  graph_obj,
  r = 0.15,
  node_add = 8L,
  orientation = c("up", "down", "left", "right"),
  angle = 0,
  y_squash = 1
)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}

\item{r}{Numeric (default = 1).
Radius increment for concentric or layered layouts.}

\item{node_add}{Integer (default = 7).
Number of nodes to add in each layer of the layout.}

\item{orientation}{Character string.
custom orientation; one of
"up","down","left","right"}

\item{angle}{Integer  (default = 0).
change  orientation angle}

\item{y_squash}{Integer   (default = 1).}
}
\value{
Layout
}
\description{
internal function
}
\examples{
NULL
}
\keyword{internal}
