% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layout_petal.R
\name{create_layout_petal}
\alias{create_layout_petal}
\title{Create petal layout}
\usage{
create_layout_petal(graph_obj, node_add = 7, r = 0.1, petals = 6, amp = 0.35)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}

\item{node_add}{Integer (default = 7).
Number of nodes to add in each layer of the layout.}

\item{r}{Numeric (default = 1).
Radius increment for concentric or layered layouts.}

\item{petals}{Numeric (default = 6).}

\item{amp}{Numeric (default = 0.35).}
}
\value{
Layout
}
\description{
internal function
}
\examples{
NULL
}
\keyword{internal}
