% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layout_petal2.R
\name{create_layout_petal2}
\alias{create_layout_petal2}
\title{Create petal2 layout}
\usage{
create_layout_petal2(
  graph_obj,
  node_add = 7,
  r = 0.1,
  petals = 6,
  amp = 0.35,
  inner_rings = 2,
  transition_rings = 0,
  orientation = c("up", "down", "left", "right"),
  angle = 0
)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}

\item{node_add}{Integer (default = 7).
Number of nodes to add in each layer of the layout.}

\item{r}{Numeric (default = 1).
Radius increment for concentric or layered layouts.}

\item{petals}{Numeric (default = 6).}

\item{amp}{Numeric (default = 0.35).}

\item{inner_rings}{Numeric (default = 2).}

\item{transition_rings}{Numeric (default = 0).}
}
\value{
Layout
}
\description{
internal function
}
\examples{
NULL
}
\keyword{internal}
