% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layout_rectangle.R
\name{create_layout_rectangle}
\alias{create_layout_rectangle}
\title{Create rectangle layout}
\usage{
create_layout_rectangle(
  graph_obj,
  node_add = 7,
  r = 0.1,
  ratio = 1.5,
  orientation = c("up", "down", "left", "right"),
  angle = 0
)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}

\item{node_add}{Integer (default = 7).
Number of nodes to add in each layer of the layout.}

\item{r}{Numeric (default = 1).
Radius increment for concentric or layered layouts.}

\item{ratio}{Numeric (default = 1.5).}

\item{orientation}{Character string.
custom orientation; one of
"up","down","left","right"}

\item{angle}{Integer  (default = 0).
change  orientation angle}
}
\value{
Layout
}
\description{
internal function
}
\examples{
NULL
}
\keyword{internal}
