% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_from_graph.R
\name{get_info_from_graph}
\alias{get_info_from_graph}
\title{Extract node and edge information from graph object}
\usage{
get_info_from_graph(graph_obj)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}
}
\value{
a list
}
\description{
Extract node and edge information from graph object
}
\examples{
data(adjacency_matrix_example)
set.seed(1)
idx <- sample(ncol(adjacency_matrix_example), 80)
obj <- build_graph_from_adj_mat(adjacency_matrix_example[idx, idx])
info <- get_info_from_graph(obj)
names(info)
head(info$node_info)
head(info$edge_info)
}
