% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_subgraph_topology.R
\name{get_sample_subgraph_topology}
\alias{get_sample_subgraph_topology}
\title{Create network topology}
\usage{
get_sample_subgraph_topology(
  graph_obj,
  mat = NULL,
  transfrom.method = c("none", "scale", "center", "log2", "log10", "ln", "rrarefy",
    "rrarefy_relative"),
  r.threshold = 0.7,
  p.threshold = 0.05,
  method = c("WGCNA", "SpiecEasi", "SPARCC", "cor"),
  cor.method = c("pearson", "kendall", "spearman"),
  proc = c("Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BH", "BY", "ABH",
    "TSBH"),
  bootstrap = 100
)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}

\item{mat}{Numeric Matrix
The matrix to build graph_obj}

\item{bootstrap}{Numeric  (default = 100).
Number of bootstrap iterations for stability analysis}
}
\value{
data frame of network topolog
}
\description{
Generates a data frame containing network topology information from either a
pre-built graph object or directly from an adjacency matrix.
}
\examples{
NULL
}
