% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subgraph.R
\name{get_subgraph}
\alias{get_subgraph}
\title{Extract subgrah from graph object}
\usage{
get_subgraph(graph_obj, select_module = NULL)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}

\item{select_module}{a character vectors
Select the module name in graph object}
}
\value{
list
}
\description{
Extract subgrah from graph object
}
\examples{
data(ppi_example)
obj <- build_graph_from_df(
  df              = ppi_example$ppi,
  node_annotation = ppi_example$annotation
)
sg <- get_subgraph(obj, select_module = "1")
names(sg)
sg$stat_module
}
