% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{ggNetView-internal}
\alias{ggNetView-internal}
\title{ggNetView internal: namespace imports}
\description{
Internal scaffold to declare base R and third-party function imports.
Not intended to be called directly.
}
\keyword{internal}
