% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_rare}
\alias{otu_rare}
\title{Rarefied OTU abundance table}
\format{
A `data.frame` with 2,859 rows (OTUs) and 18 columns (samples).
}
\source{
Derived from [otu_tab] by rarefaction.
}
\usage{
otu_rare
}
\description{
A rarefied version of [otu_tab] in which every sample has been
subsampled to a common sequencing depth.
}
\examples{
data(otu_rare)
otu_rare[1:5, 1:5]
}
\keyword{datasets}
