% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_rare_relative}
\alias{otu_rare_relative}
\title{Relative-abundance OTU table}
\format{
A `data.frame` with 2,859 rows and 18 columns.
}
\source{
Derived from [otu_rare].
}
\usage{
otu_rare_relative
}
\description{
A relative-abundance transformation of [otu_rare]: each column sums
approximately to 1.
}
\examples{
data(otu_rare_relative)
otu_rare_relative[1:5, 1:5]
}
\keyword{datasets}
