% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_sample}
\alias{otu_sample}
\title{Sample metadata for the OTU example}
\format{
A `data.frame` with 18 rows (samples) and 2 columns.
}
\source{
Internal example data shipped with ggNetView.
}
\usage{
otu_sample
}
\description{
Sample-level metadata accompanying [otu_tab].
}
\examples{
data(otu_sample)
head(otu_sample)
}
\keyword{datasets}
