% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppi_example}
\alias{ppi_example}
\title{Protein-protein interaction example}
\format{
A list with two elements:
\describe{
  \item{`ppi`}{Edge `data.frame` (`from`, `to`, `weight`).}
  \item{`annotation`}{Node annotation `data.frame`.}
}
}
\source{
Subset of the STRING database (https://string-db.org/).
}
\usage{
ppi_example
}
\description{
A small protein-protein interaction (PPI) network used by the
`build_graph_from_pie` and STRING-DB examples. The list contains
the interaction edge table and a node annotation table.
}
\examples{
data(ppi_example)
names(ppi_example)
}
\keyword{datasets}
