% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spieceasi_matrix_rcpp.R
\name{spieceasi_fit_rcpp}
\alias{spieceasi_fit_rcpp}
\title{SPIEC-EASI fit (Rcpp-accelerated CLR)}
\usage{
spieceasi_fit_rcpp(
  data,
  method = "glasso",
  sel.criterion = "stars",
  verbose = TRUE,
  pulsar.select = TRUE,
  pulsar.params = list(),
  stars.params = list(),
  lambda.log = TRUE,
  ...
)
}
\description{
Fits the SpiecEasi model using fast C++ CLR transformation.
}
\keyword{internal}
