% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax_NatureWater}
\alias{tax_NatureWater}
\title{Taxonomy table for [otu_NatureWater]}
\format{
A `tbl_df` with 51,998 rows and 18 columns.
}
\source{
Adapted from the same published study as [otu_NatureWater].
}
\usage{
tax_NatureWater
}
\description{
Taxonomic annotation matching [otu_NatureWater].
}
\examples{
data(tax_NatureWater)
head(tax_NatureWater)
}
\keyword{datasets}
