% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph_from_mat.R
\name{build_graph_from_mat}
\alias{build_graph_from_mat}
\title{Build a graph object from raw matrix for correlation analysis}
\usage{
build_graph_from_mat(
  mat,
  r.threshold = 0.7,
  p.threshold = 0.05,
  method = c("WGCNA", "SpiecEasi", "SPARCC"),
  cor.method = c("pearson", "kendall", "spearman"),
  SpiecEasi.method = c("mb", "glasso"),
  proc = c("Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BH", "BY", "ABH",
    "TSBH"),
  module.method = c("Fast_greedy", "Walktrap", "Edge_betweenness", "Spinglass"),
  node_annotation = NULL,
  top_modules = 15,
  seed = 1115
)
}
\arguments{
\item{mat}{Numeric matrix.
A matrix with variables in columns; pairwise correlations are computed.}

\item{r.threshold}{Numeric.
Correlation coefficient threshold; edges are kept only if |r| >= r.threshold.}

\item{p.threshold}{Numeric.
Significance threshold for correlations; edges are kept only if p < p.threshold.}

\item{method}{Charecter.
Relationship analysis methods contains "WGCNA", "SpiecEasi", "SPARCC"}

\item{cor.method}{Charecter.
Correlation analysis methods contains "pearson", "kendall", "spearman"}

\item{SpiecEasi.method}{Charecter.
SpiecEasi methods contains "mb", "glasso"}

\item{proc}{Corelation adjust pvalue methods contains "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD","BH", "BY","ABH","TSBH"}

\item{module.method}{Character
Module analysis methods contains "Fast_greedy", "Walktrap", "Edge_betweenness", "Spinglass"}

\item{node_annotation}{Data Frame
The annotation file of nodes in network}

\item{top_modules}{Integer.
Number of top-ranked modules to retain}

\item{seed}{Integer.
Random seed for reproducibility; if `NULL`, no seed is set.}
}
\value{
An graph object representing the correlation network.
Node/edge attributes include correlation statistics and (optionally) module labels.
}
\description{
Build a graph object from raw matrix for correlation analysis
}
\examples{
NULL
}
