% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ggnetview.R
\name{export_ggnetview}
\alias{export_ggnetview}
\title{Save a ggNetView with sensible defaults}
\usage{
export_ggnetview(p, filename, height, width, limitsize = F, dpi = 300)
}
\arguments{
\item{p}{Plot to save, defaults to last plot displayed.}

\item{filename}{File name to create on disk.}

\item{height}{Plot size in units expressed by the units argument. If not supplied, uses the size of the current graphics device.}

\item{width}{Plot size in units expressed by the units argument. If not supplied, uses the size of the current graphics device.}

\item{limitsize}{When TRUE (the default), export_ggnetview() will not save images larger than 50x50 inches, to prevent the common error of specifying dimensions in pixels.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Only applies when converting pixel units, as is typical for raster output types.}
}
\value{
a graph output
}
\description{
Save a ggNetView with sensible defaults
}
\examples{
\donttest{
data(ppi_example)
obj <- build_graph_from_df(
  df              = ppi_example$ppi,
  node_annotation = ppi_example$annotation
)
p <- ggNetView(obj, layout = "fr", layout.module = "adjacent")
tmp <- tempfile(fileext = ".png")
export_ggnetview(p, filename = tmp, height = 6, width = 6)
file.exists(tmp)
}
}
