% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_topology.R
\name{get_network_topology}
\alias{get_network_topology}
\title{Create network topolog}
\usage{
get_network_topology(graph_obj)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}
}
\value{
data frame of network topolog
}
\description{
Create network topolog
}
\examples{
NULL
}
