% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_topology.R
\name{get_network_topology}
\alias{get_network_topology}
\title{Create network topology}
\usage{
get_network_topology(
  graph_obj,
  mat = NULL,
  transfrom.method = c("none", "scale", "center", "log2", "log10", "ln", "rrarefy",
    "rrarefy_relative"),
  r.threshold = 0.7,
  p.threshold = 0.05,
  method = c("WGCNA", "SpiecEasi", "SPARCC", "cor"),
  cor.method = c("pearson", "kendall", "spearman"),
  proc = c("Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BH", "BY", "ABH",
    "TSBH"),
  SpiecEasi.method = c("mb", "glasso"),
  sparcc_R = 20,
  bootstrap = 100
)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}

\item{mat}{Numeric Matrix (default = NULL)
The matrix to build graph_obj}

\item{transfrom.method}{Character.
Data transformation methods applied before correlation analysis.
Options include:
"none" (raw data),
"scale" (z-score standardization),
"center" (mean centering only),
"log2" (log2 transfrom),
"log10" (log10 transfrom),
"ln" (natural transfrom ),
"rrarefy" (random rarefaction using \code{vegan::rrarefy}),
"rrarefy_relative" (rarefy then convert to relative abundance).}

\item{r.threshold}{Numeric.
Correlation coefficient threshold; edges are kept only if |r| >= r.threshold.}

\item{p.threshold}{Numeric.
Significance threshold for correlations; edges are kept only if p < p.threshold.}

\item{method}{Character.
Relationship analysis methods.
Options include: "WGCNA", "SpiecEasi", "SPARCC" and "cor".}

\item{cor.method}{Character.
Correlation analysis method.
Options include "pearson", "kendall", and "spearman".}

\item{proc}{Character.
Correlation p-value adjustment methods.
Options include:
"Bonferroni", "Holm", "Hochberg", "
SidakSS", "SidakSD","BH",
"BY", "ABH", and "TSBH".}

\item{sparcc_R}{Integer.
Number of bootstrap/permutation replicates for SparCC p-values (when \code{method = "SPARCC"}).
Default 20.}

\item{bootstrap}{Numeric  (default = 100).
Number of bootstrap iterations for stability analysis}
}
\value{
data frame of network topolog
}
\description{
Generates a data frame containing network topology information from either a
pre-built graph object or directly from an adjacency matrix.
}
\examples{
NULL
}
