% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subgraph.R
\name{get_subgraph}
\alias{get_subgraph}
\title{Extract subgrah from graph object}
\usage{
get_subgraph(graph_obj)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}
}
\value{
list
}
\description{
Extract subgrah from graph object
}
\examples{
NULL
}
