% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglink_heatmap_triple.R
\name{gglink_heatmap_triple}
\alias{gglink_heatmap_triple}
\title{Visualize multi-orientation environmental–species correlation heatmaps2}
\usage{
gglink_heatmap_triple(Environment, Experiment, edge, node, r = 6)
}
\arguments{
\item{Environment}{character
The file Path of Environment data}

\item{Experiment}{character
The file Path of Experiment data}

\item{edge}{character
The file Path of edge data}

\item{node}{character
The file Path of node data}

\item{r}{numeric (default = 6)}
}
\value{
a ggplot2 object
}
\description{
Visualize multi-orientation environmental–species correlation heatmaps2
}
\examples{
NULL
}
