% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetview.R
\name{ggNetView}
\alias{ggNetView}
\title{Visualize a network with custom layouts}
\usage{
ggNetView(
  graph_obj,
  layout = NULL,
  node_add = 7,
  r = 1,
  center = T,
  idx = NULL,
  shrink = 1,
  label = F,
  add_outer = F
)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}

\item{layout}{Character string.
custom layouts; one of
"gephi", "square", "square2", "petal",
"petal2", "heart_centered","diamond",
"star", "star_concentric","rectangle,
"rightiso_layers"}

\item{node_add}{Integer (default = 7).
Number of nodes to add in each layer of the layout.}

\item{r}{Numeric (default = 1).
Radius increment for concentric or layered layouts.}

\item{center}{Logical (default = TRUE).
Whether to place a node at the center of the layout.}

\item{idx}{Optional.
Index of nodes to be emphasized or centered in the layout}

\item{shrink}{Numeric (default = 1).
Shrinkage factor applied to the center points.}

\item{label}{Logical (default = FALSE).
Whether to display node labels in the center points}

\item{add_outer}{Logical (default = FALSE).
Whether to add an outer circle/border around the layout.}
}
\value{
A ggplot object representing the network visualization.
}
\description{
Visualize a network with custom layouts
}
\examples{
NULL
}
