% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetview.R
\name{ggNetView}
\alias{ggNetView}
\title{Visualize a network with custom layouts}
\usage{
ggNetView(
  graph_obj,
  layout = NULL,
  node_add = 7,
  ring_n = NULL,
  r = 1,
  center = T,
  idx = NULL,
  shrink = 1,
  split = 1,
  group.by = NULL,
  label = F,
  nodelabsize = 5,
  labelsize = 10,
  labelsegmentsize = 1,
  labelsegmentalpha = 1,
  linealpha = 0.25,
  linecolor = "grey70",
  add_outer = F,
  remove = F,
  outerwidth = 1.25,
  outerlinetype = 2,
  outeralpha = 0.5,
  orientation = "up",
  angle = 0
)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}

\item{layout}{Character string.
custom layouts; one of
"gephi", "square", "square2", "petal",
"petal2", "heart_centered","diamond",
"star", "star_concentric","rectangle,
"rightiso_layers"}

\item{node_add}{Integer (default = 7).
Number of nodes to add in each layer of the layout.}

\item{ring_n}{Integer  (default = NULL).
change  ring number in multipy ring layout}

\item{r}{Numeric (default = 1).
Radius increment for concentric or layered layouts.}

\item{center}{Logical (default = TRUE).
Whether to place a node at the center of the layout.}

\item{idx}{Optional.
Index of nodes to be emphasized or centered in the layout}

\item{shrink}{Numeric (default = 1).
Shrinkage factor applied to the center points.}

\item{split}{Numeric (default = 1).
split factor applied to the center points.}

\item{group.by}{Character   (default = "Modularity").
change variable for group}

\item{label}{Logical (default = FALSE).
Whether to display node labels in the center points}

\item{nodelabsize}{Integer  (default = 5).
change  node label size}

\item{labelsize}{Integer  (default = 10).
change  label size}

\item{labelsegmentsize}{Integer  (default = 1).
change  label segment size}

\item{labelsegmentalpha}{Integer  (default = 1).
change  label segment alpha}

\item{linealpha}{Integer  (default = 0.25).
change  line alpha}

\item{linecolor}{Character  (default = "grey70").
change  line color}

\item{add_outer}{Logical (default = FALSE).
Whether to add an outer circle/border around the layout.}

\item{remove}{Logical (default = FALSE).
delect nodes that are not modules}

\item{outerwidth}{Integer  (default = 1.25).
change  outer linewidth}

\item{outerlinetype}{Integer  (default = 2).
change  outer linetype}

\item{outeralpha}{Integer  (default = 0.5).
change  outer alpha}

\item{orientation}{Character string.
custom orientation; one of
"up","down","left","right"}

\item{angle}{Integer  (default = 0).
change  orientation angle}
}
\value{
A ggplot object representing the network visualization.
}
\description{
Visualize a network with custom layouts
}
\examples{
NULL
}
