% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggNetView_multi_link.R
\name{ggNetView_multi_link}
\alias{ggNetView_multi_link}
\title{Visualize multiple networks and show the connections between them to highlight key members within the networks}
\usage{
ggNetView_multi_link(
  mat,
  group_info,
  transfrom.method = c("none", "scale", "center", "log2", "log10", "ln", "rrarefy",
    "rrarefy_relative"),
  r.threshold = 0.7,
  p.threshold = 0.05,
  method = c("WGCNA", "SpiecEasi", "SPARCC", "cor"),
  cor.method = c("pearson", "kendall", "spearman"),
  proc = c("Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BH", "BY", "ABH",
    "TSBH"),
  module.method = c("Fast_greedy", "Walktrap", "Edge_betweenness", "Spinglass"),
  SpiecEasi.method = c("mb", "glasso"),
  node_annotation = NULL,
  top_modules = 15,
  layout = NULL,
  node_add = 7,
  ring_n = NULL,
  r = 1,
  center = TRUE,
  idx = NULL,
  shrink = 1,
  k_nn = 12,
  push_others_delta = 0,
  layout.module = c("random", "adjacent", "order"),
  group.by = "Modularity",
  fill.by = "Modularity",
  jitter = FALSE,
  jitter_sd = 0.01,
  mapping_line = FALSE,
  linealpha = 0.25,
  linecolor = "grey70",
  scale = T,
  orientation = "up",
  angle = 0,
  anchor_dist = 6,
  seed = 1115
)
}
\arguments{
\item{mat}{Numeric matrix.
A numeric matrix with samples in rows and variables in columns.}

\item{group_info}{DataFrame
The group information contains: Sample and Group}

\item{transfrom.method}{Character.
Data transformation methods applied before correlation analysis.
Options include:
"none" (raw data),
"scale" (z-score standardization),
"center" (mean centering only),
"log2" (log2 transfrom),
"log10" (log10 transfrom),
"ln" (natural transfrom ),
"rrarefy" (random rarefaction using \code{vegan::rrarefy}),
"rrarefy_relative" (rarefy then convert to relative abundance).}

\item{r.threshold}{Numeric.
Correlation coefficient threshold; edges are kept only if |r| >= r.threshold.}

\item{p.threshold}{Numeric.
#' Significance threshold for correlations; edges are kept only if p < p.threshold.}

\item{method}{Character.
Relationship analysis methods.
Options include: "WGCNA", "SpiecEasi", "SPARCC" and "cor".}

\item{cor.method}{Character.
Correlation analysis method.
Options include "pearson", "kendall", and "spearman".}

\item{proc}{Character.
Correlation p-value adjustment methods.
Options include:
"Bonferroni", "Holm", "Hochberg", "
SidakSS", "SidakSD","BH",
"BY", "ABH", and "TSBH".}

\item{module.method}{Character.
Network community detection (module identification) method.
Options include "Fast_greedy", "Walktrap", "Edge_betweenness", and "Spinglass".}

\item{SpiecEasi.method}{Character.
Method used in \code{SpiecEasi} network inference; options include "mb" and "glasso".}

\item{node_annotation}{Data frame.
Optional node annotation table, containing metadata such as taxonomy or functional categories.}

\item{top_modules}{Integer.
Number of top-ranked modules to retain for downstream visualization or analysis.}

\item{layout}{Character string.
Custom layouts; one of "gephi", "square", "square2", "petal",
"petal2", "heart_centered","diamond", "star", "star_concentric","rectangle,
"rightiso_layers" etc.}

\item{node_add}{Integer (default = 7).
Number of nodes to add in each layer of the layout.}

\item{ring_n}{Numeric (default = 7)
Numbers of ring in rings layout.}

\item{r}{Numeric (default = 1).
Radius increment for concentric or layered layouts.}

\item{center}{Logical (default = TRUE).
Whether to place a node at the center of the layout.}

\item{idx}{Optional.
Index of nodes to be emphasized or centered in the layout}

\item{shrink}{Numeric (default = 1).
Shrinkage factor applied to the center points.}

\item{k_nn}{Numeric (default = 8).
Number of nearest neighbors used to build the local adjacency graph.}

\item{push_others_delta}{Numeric (default = 0).
Radial offset applied to the "Others" module to slightly}

\item{layout.module}{Character  (default = "random")
- adjacent : modules are positioned close to each other, minimizing inter-module gaps.
- order : modules are distributed by order, applicable to `Bipartite, Tripartite, Quadripartite, Multipartite, Pentapartite Layout`}

\item{group.by}{Character (default = "Modularity").
Change group for nodes}

\item{fill.by}{Character (default = "Modularity").
Change fill for nodes}

\item{jitter}{Logical (default = FALSE).
Whether to apply jitter to points.}

\item{jitter_sd}{Integer  (default = 0.1).
The standard deviation of the jitter applied when `jitter = TRUE`.}

\item{mapping_line}{Logical (default = FALSE).
Whether to mapping line in ggNetView.}

\item{linealpha}{Integer  (default = 0.25).
Change  line alpha.}

\item{linecolor}{Character  (default = "grey70").
Change  line color.}

\item{scale}{Logical  (default = T).
modules applicable to `Bipartite, Tripartite, Quadripartite, Multipartite, Pentapartite Layout` to scale the radius}

\item{orientation}{Character string.
Custom orientation; one of "up","down","left","right".}

\item{angle}{Integer  (default = 0).
Change  orientation angle.}

\item{anchor_dist}{Integer (default = 2)
the distance of each modules, applicable to `Bipartite, Tripartite, Quadripartite, Multipartite, Pentapartite Layout`}

\item{seed}{Integer (default = 1115).
Random seed for reproducibility.}

\item{pointstroke}{Integer  (default = 0.3).}
}
\value{
A ggplot object representing the network visualization.
}
\description{
Visualize multiple networks and show the connections between them to highlight key members within the networks
}
\examples{
NULL
}
