% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_graph.R
\name{order_graph}
\alias{order_graph}
\title{Reorder the Modularity from graph object}
\usage{
order_graph(graph_obj, order)
}
\arguments{
\item{graph_obj}{An graph object from build_graph_from_mat or build_graph_from_df.
The network object to be visualized.}

\item{order}{a character vectors}
}
\value{
An graph object with reorder
}
\description{
Reorder the Modularity from graph object
}
\examples{
NULL
}
