% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_adjacency_matrix_to_df.R
\name{trans_adjacency_matrix_to_df}
\alias{trans_adjacency_matrix_to_df}
\title{Trans adjacency matrix to edge data frame}
\usage{
trans_adjacency_matrix_to_df(adjacency_matrix)
}
\arguments{
\item{adjacency_matrix}{Numeric matrix.
A numeric matrix with adjacency matrix.}
}
\value{
a data frame of edge
}
\description{
Trans adjacency matrix to edge data frame
}
\examples{
data(adjacency_matrix_example)
set.seed(1)
idx <- sample(ncol(adjacency_matrix_example), 50)
edge_df <- trans_adjacency_matrix_to_df(adjacency_matrix_example[idx, idx])
head(edge_df)
}
